+++ examples/spi_dma.rs
@@ -19,6 +19,7 @@ fn main() -> ! {
     let p = pac::Peripherals::take().unwrap();
 
     let mut rcc = p.RCC.constrain();
+    let clocks = rcc.cfgr.freeze();
 
     let dma = DMA::new(p.DMA1);
     let gpiob = p.GPIOB.split();
@@ -46,12 +47,13 @@ fn main() -> ! {
 
     // Initialize SPI
     let mut spi = Spi::new(p.SPI3, (sck, miso, mosi)).enable(
-        &mut rcc.apb1,
-        spi::ClockDivider::DIV32,
         spi::Mode {
             polarity: spi::Polarity::IdleHigh,
             phase: spi::Phase::CaptureOnSecondTransition,
         },
+        250.kHz(),
+        &clocks,
+        &mut rcc.apb1,
     );
 
     // Create the buffer we're going to use for DMA. This is safe, as this